feat: 좌석 선택 및 해제 이벤트 발생 시 사용자 수 브로드 캐스트 로직 구현 - 동시성 제어 필요

+++ src/main/java/com/moti/backend/core/show/presentation/socket/dto/SeatResponseDTO.java
@@ -18,11 +18,11 @@ public class SeatResponseDTO {
 		private final Long showScheduleId;
 		private final Long seatId;
 		private final SeatStatus currentStatus;
-		private final int selectedCount;
+		private final Long selectedCount;
 		private final LocalDateTime timestamp;
 
 		public static SeatToggleResponse from(EventType eventType, Long showScheduleId, Long seatId,
-			SeatStatus currentStatus, int selectedCount, LocalDateTime timestamp) {
+			SeatStatus currentStatus, Long selectedCount, LocalDateTime timestamp) {
 			return SeatToggleResponse.builder()
 				.eventType(eventType)
 				.showScheduleId(showScheduleId)
